requirements.txt
fastapi 
uvicorn 
google-generativeai 
python-multipart 
pandas
python-dotenv